Update LLVM to 3.1, this breaks backwards compatibility.
Updated the code base to use the llvm 3.1 release instead of the previous version. This breaks backwards compatibility with llvm 2.9. I have not tested with llvm 3.0, but it might work. The major differences are in the execution handling code gen, the introduction of the ArrayRef<> and a number of data types are no longer const in the llvm API.

It compiles and runs, but some specs are still failing. Still investigating the cause of the failed specs.